Rendering-hooks are now registered in a registry separate from the MapRenderer. Added a BackgroundLoadingTask to reduce boilerplate in such loading-tasks.

+++ core/src/main/groovy/org/snowjak/city/screens/loadingtasks/AssetServiceLoadingTask.java
@@ -32,7 +32,7 @@ public class AssetServiceLoadingTask extends LoadingTask {
 	}
 	
 	@Override
-	public float getProgress() {
+	public double getProgress() {
 		
 		assetService.update();
 		return assetService.getProgress();
